refactor: consolidate diff helpers and move WritePhase to output.rs
- Move render_diffs_plain() and render_diffs_colored() to diff.rs as
  pub(crate) helpers, removing 3 duplicate copies from output.rs,
  replace.rs, and tidy.rs.

- Move WritePhase enum from write_dispatch.rs to output.rs (its
  natural home post-migration). Update all imports across append,
  create, delete, doc, md, rename. write_dispatch.rs re-imports
  from output.rs for its remaining execute_write() function.
